Added the `ValidatorGenerator` class, which, when run, will generate an extension / category that allows the core data object that the validator was generated for to be validated by running `[object validate]` / `object.validate()` depending on the language.
These validators currently only support presence validation in Rails, but the fields are pulled from the Rails models which come with the validators, so support for new validations shouldn't be much of a drama.

Added the swift template so far, which is just a copy of ones I've written already, doing the Obj-C version next.