Major restructure of how the `validation` categories are generated.
Instead of creating an instance of `PCDValidator` directly, they now create an instance of `RKGValidator`, which is our wrapper to any choice of validation library the developer might make. As long as the appropriate methods, `addPresenceValidatorForField` and `hasErrors`, are defined, the categories can use that to dictate where an object instance is valid or not.

It's expected that `hasErrors` sets the `invalidKeyPaths` attribute of the `RKGValidator` instance, and using `objc_setAssociatedObject`, we can add the keyPaths as a property of our object, which can then be grabbed post-validation if it fails.

The `errorMessage:` method has been changed to a class method which can just be passed a keypath, and the appropriate error message is returning if defined (we leave it up to the developers to define those localised keys themselves).