Adds support for a `drop_namespace` option for models.
Basically, if I want to generate a model that only exists within a gem, i.e. `Magnetik::CreditCard`, I want an iOS model called `CreditCard`, not `MagnetikCreditCard`. Sure, I could just define the model in the Rails app, but it's nice to have to have the option to not have to do that.