#37 - Quick spec and dummy app additions to reflect the problem being addressed.
A named route for a member action, i.e. `report_api_post`, currently generates the category named `ReportApiPostRoute`, not removing the namepsace from the route because it's not the prefix of the route name. We want the `ios_route_name` to be `report_post`, and the category to ultimately be `ReportPostRoute`.