#37 - Implements a new namespace stripping strategy, to remove a namespace even if it sits in front of a member action.
Basically, a member actions name will sit as the prefix to a named route, so we can't just check if the prefix of a named route matches our specified namespace. Instead, we check if the section after a member action (if there is one) matches the namespace, and we just hard slice out the section and return the remainder (which also works for the normal cases where the namespace _is_ the prefix.

Could've gone for a gsub, but I didn't want such a blanket solution in case there was multiple references to the namespace in the route name. Saying that, not 100% happy with the current solution, think it can be cleaner, but it passes tests, so yay.