Adds support for an `honour_parent` option when taking parent models into account.
Had an issue where the parent class of an object wasn't being generated into the app because it lived in a gem and the the subclass basically just extended it for a specific project; however generating the child excluded all the fields the parent had because it was a subclass. So this option is a way of just saying "no, I still want all the stuff in the parent class.